requirements.txt
torch==1.7.1
Flask==1.1.1
Jinja2==2.10.1
gunicorn==19.9.0
certifi==2019.11.28
itsdangerous==1.1.0
opencv-python==4.5.1.48
Pillow==2.2.1
